feat, wip: least frequently updated cache
This commit adds the Insert and Access methods to
LFU cache. "Frequency nodes" currenly rely on a map
to keep registry of nodes seen a number of times
the value it holds. It is ok, but the method to evict
the LRU node of a LFU frequency node is NOT deterministic
since maps does not remember the order of items. Thus,
consider this a WIP.